+++ bgpd/bgp_vty.h
@@ -186,6 +186,6 @@ int bgp_vty_find_and_parse_bgp(struct vty *vty, struct cmd_token **argv,
 			       int argc, struct bgp **bgp, bool use_json);
 extern int bgp_show_summary_vty(struct vty *vty, const char *name, afi_t afi,
 				safi_t safi, const char *neighbor, int as_type,
-				as_t as, uint8_t show_flags);
+				as_t as, uint16_t show_flags);
 
 #endif /* _QUAGGA_BGP_VTY_H */
